+++ src/Exception/SignatureInvalidException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace achertovsky\jwt\Exception;
+
+class SignatureInvalidException extends JwtException
+{
+}